requirements.txt
backcall==0.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
fonttools==4.38.0
idna==3.10
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.1
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.21.6
packaging==24.0
pandas==1.3.5
parso==0.8.4
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
prompt_toolkit==3.0.47
psutil==6.0.0
Pygments==2.17.2
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
pyzmq==26.2.0
requests==2.31.0
scipy==1.7.3
six==1.16.0
sv-ttk==2.5.5
tk==0.1.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
yarg==0.1.10
